// src/types/RestCountryApiResponse.ts
export type RestCountryApiResponse = RestCountry[]

export interface RestCountry {
  name: Name
  tld?: string[]
  cca2: string
  ccn3?: string
  cca3: string
  cioc?: string
  independent?: boolean
  status: string
  unMember: boolean
  currencies?: Currencies
  idd: Idd
  capital?: string[]
  altSpellings: string[]
  region: string
  subregion?: string
  languages?: Languages
  translations: Translations
  latlng: number[]
  landlocked: boolean
  borders?: string[]
  area: number
  demonyms?: Demonyms
  flag?: string
  maps: Maps
  population: number
  gini?: Gini
  fifa?: string
  car: Car
  timezones: string[]
  continents: string[]
  flags: Flags
  coatOfArms: CoatOfArms
  startOfWeek: string
  capitalInfo: CapitalInfo
  postalCode?: PostalCode
}

interface Name {
  common: string
  official: string
  nativeName?: NativeName
}

interface NativeName {
  eng?: nativeType
  swa?: nativeType
  fra?: nativeType
  por?: nativeType
  pov?: nativeType
  hmo?: nativeType
  tpi?: nativeType
  spa?: nativeType
  kal?: nativeType
  sag?: nativeType
  msa?: nativeType
  crs?: nativeType
  ind?: nativeType
  ara?: nativeType
  deu?: nativeType
  ltz?: nativeType
  zho?: nativeType
  tam?: nativeType
  div?: nativeType
  bwg?: nativeType
  kck?: nativeType
  khi?: nativeType
  ndc?: nativeType
  nde?: nativeType
  nya?: nativeType
  sna?: nativeType
  sot?: nativeType
  toi?: nativeType
  tsn?: nativeType
  tso?: nativeType
  ven?: nativeType
  xho?: nativeType
  zib?: nativeType
  fin?: nativeType
  swe?: nativeType
  kor?: nativeType
  pih?: nativeType
  lao?: nativeType
  kon?: nativeType
  lin?: nativeType
  bul?: nativeType
  gsw?: nativeType
  ita?: nativeType
  roh?: nativeType
  jpn?: nativeType
  dan?: nativeType
  nld?: nativeType
  pap?: nativeType
  rus?: nativeType
  uzb?: nativeType
  cnr?: nativeType
  ssw?: nativeType
  vie?: nativeType
  tuk?: nativeType
  som?: nativeType
  sin?: nativeType
  bel?: nativeType
  est?: nativeType
  ron?: nativeType
  fij?: nativeType
  hif?: nativeType
  sqi?: nativeType
  hun?: nativeType
  slv?: nativeType
  tur?: nativeType
  mlt?: nativeType
  mya?: nativeType
  ces?: nativeType
  slk?: nativeType
  fil?: nativeType
  gle?: nativeType
  smo?: nativeType
  tkl?: nativeType
  urd?: nativeType
  hye?: nativeType
  mri?: nativeType
  nzs?: nativeType
  tvl?: nativeType
  tha?: nativeType
  aym?: nativeType
  que?: nativeType
  ell?: nativeType
  mlg?: nativeType
  grn?: nativeType
  nor?: nativeType
  lat?: nativeType
  bos?: nativeType
  hrv?: nativeType
  srp?: nativeType
  bis?: nativeType
  hat?: nativeType
  tir?: nativeType
  gil?: nativeType
  ton?: nativeType
  tgk?: nativeType
  prs?: nativeType
  pus?: nativeType
  jam?: nativeType
  lav?: nativeType
  kaz?: nativeType
  glv?: nativeType
  ber?: nativeType
  mey?: nativeType
  kin?: nativeType
  nfr?: nativeType
  lit?: nativeType
  nno?: nativeType
  nob?: nativeType
  smi?: nativeType
  dzo?: nativeType
  arc?: nativeType
  ckb?: nativeType
  kir?: nativeType
  hin?: nativeType
  nrf?: nativeType
  nep?: nativeType
  aze?: nativeType
  amh?: nativeType
  niu?: nativeType
  pol?: nativeType
  zdj?: nativeType
  run?: nativeType
  cat?: nativeType
  rar?: nativeType
  fas?: nativeType
  mkd?: nativeType
  khm?: nativeType
  afr?: nativeType
  her?: nativeType
  hgm?: nativeType
  kwn?: nativeType
  loz?: nativeType
  ndo?: nativeType
  isl?: nativeType
  heb?: nativeType
  nbl?: nativeType
  nso?: nativeType
  zul?: nativeType
  ben?: nativeType
  lua?: nativeType
  ukr?: nativeType
  cal?: nativeType
  cha?: nativeType
  pau?: nativeType
  fao?: nativeType
  bar?: nativeType
  kat?: nativeType
  bjz?: nativeType
  mfe?: nativeType
  mah?: nativeType
  tet?: nativeType
  mon?: nativeType
  nau?: nativeType
}

export interface nativeType {
  official: string
  common: string
}

interface Currencies {
  TZS?: currencyType
  EUR?: currencyType
  XOF?: currencyType
  FKP?: currencyType
  PGK?: currencyType
  COP?: currencyType
  DKK?: currencyType
  XAF?: currencyType
  SLL?: currencyType
  GBP?: currencyType
  SHP?: currencyType
  MXN?: currencyType
  BND?: currencyType
  SGD?: currencyType
  SCR?: currencyType
  IDR?: currencyType
  LYD?: currencyType
  GNF?: currencyType
  MYR?: currencyType
  AOA?: currencyType
  MVR?: currencyType
  ZWL?: currencyType
  TND?: currencyType
  KRW?: currencyType
  AUD?: currencyType
  LAK?: currencyType
  JOD?: currencyType
  BGN?: currencyType
  HKD?: currencyType
  CHF?: currencyType
  CVE?: currencyType
  JPY?: currencyType
  AWG?: currencyType
  UZS?: currencyType
  XCD?: currencyType
  ZMW?: currencyType
  SZL?: currencyType
  ZAR?: currencyType
  VND?: currencyType
  CRC?: currencyType
  TMT?: currencyType
  EGP?: currencyType
  ILS?: currencyType
  USD?: currencyType
  SOS?: currencyType
  NIO?: currencyType
  BMD?: currencyType
  LKR?: currencyType
  BYN?: currencyType
  MDL?: currencyType
  WST?: currencyType
  NOK?: currencyType
  BAM?: currencyType
  VUV?: currencyType
  HTG?: currencyType
  ERN?: currencyType
  KID?: currencyType
  TOP?: currencyType
  TJS?: currencyType
  RSD?: currencyType
  AFN?: currencyType
  JMD?: currencyType
  SBD?: currencyType
  XPF?: currencyType
  SDG?: currencyType
  VES?: currencyType
  KZT?: currencyType
  BWP?: currencyType
  IMP?: currencyType
  DZD?: currencyType
  MAD?: currencyType
  MRU?: currencyType
  CAD?: currencyType
  SYP?: currencyType
  RWF?: currencyType
  GGP?: currencyType
  MOP?: currencyType
  SEK?: currencyType
  KES?: currencyType
  BSD?: currencyType
  BTN?: currencyType
  INR?: currencyType
  CUC?: currencyType
  CUP?: currencyType
  IQD?: currencyType
  SSP?: currencyType
  KGS?: currencyType
  JEP?: currencyType
  BBD?: currencyType
  NPR?: currencyType
  AZN?: currencyType
  ETB?: currencyType
  PLN?: currencyType
  NGN?: currencyType
  KMF?: currencyType
  TTD?: currencyType
  BIF?: currencyType
  CKD?: currencyType
  PAB?: currencyType
  IRR?: currencyType
  MKD?: currencyType
  KHR?: currencyType
  HNL?: currencyType
  KWD?: currencyType
  OMR?: currencyType
  GIP?: currencyType
  NAD?: currencyType
  ANG?: currencyType
  ISK?: currencyType
  HRK?: currencyType
  KYD?: currencyType
  GYD?: currencyType
  GTQ?: currencyType
  BDT?: currencyType
  RUB?: currencyType
  CDF?: currencyType
  LSL?: currencyType
  CNY?: currencyType
  UAH?: currencyType
  FOK?: currencyType
  ARS?: currencyType
  UYU?: currencyType
  PYG?: currencyType
  GMD?: currencyType
  DJF?: currencyType
  GEL?: currencyType
  BZD?: currencyType
  MUR?: currencyType
  LBP?: currencyType
  SAR?: currencyType
  BRL?: currencyType
  UGX?: currencyType
  QAR?: currencyType
  MNT?: currencyType
  CLP?: currencyType
  LRD?: currencyType
  TWD?: currencyType
  FJD?: currencyType
  ALL?: currencyType
  AED?: currencyType
  HUF?: currencyType
  MZN?: currencyType
  TRY?: currencyType
  GHS?: currencyType
  DOP?: currencyType
  RON?: currencyType
  MMK?: currencyType
  CZK?: currencyType
  PHP?: currencyType
  MWK?: currencyType
  NZD?: currencyType
  PKR?: currencyType
  AMD?: currencyType
  YER?: currencyType
  KPW?: currencyType
  TVD?: currencyType
  THB?: currencyType
  SRD?: currencyType
  PEN?: currencyType
  BHD?: currencyType
  STN?: currencyType
  MGA?: currencyType
  BOB?: currencyType
}

export interface currencyType {
  name: string
  symbol: string
}

export interface Idd {
  root?: string
  suffixes?: string[]
}

export interface Languages {
  eng?: string
  swa?: string
  fra?: string
  por?: string
  pov?: string
  hmo?: string
  tpi?: string
  spa?: string
  kal?: string
  sag?: string
  msa?: string
  crs?: string
  ind?: string
  ara?: string
  deu?: string
  ltz?: string
  zho?: string
  tam?: string
  div?: string
  bwg?: string
  kck?: string
  khi?: string
  ndc?: string
  nde?: string
  nya?: string
  sna?: string
  sot?: string
  toi?: string
  tsn?: string
  tso?: string
  ven?: string
  xho?: string
  zib?: string
  fin?: string
  swe?: string
  kor?: string
  pih?: string
  lao?: string
  kon?: string
  lin?: string
  bul?: string
  gsw?: string
  ita?: string
  roh?: string
  jpn?: string
  dan?: string
  nld?: string
  pap?: string
  rus?: string
  uzb?: string
  cnr?: string
  ssw?: string
  vie?: string
  tuk?: string
  som?: string
  sin?: string
  bel?: string
  est?: string
  ron?: string
  fij?: string
  hif?: string
  sqi?: string
  hun?: string
  slv?: string
  tur?: string
  mlt?: string
  mya?: string
  ces?: string
  slk?: string
  fil?: string
  gle?: string
  smo?: string
  tkl?: string
  urd?: string
  hye?: string
  mri?: string
  nzs?: string
  tvl?: string
  tha?: string
  aym?: string
  que?: string
  ell?: string
  mlg?: string
  grn?: string
  nor?: string
  lat?: string
  bos?: string
  hrv?: string
  srp?: string
  bis?: string
  hat?: string
  tir?: string
  gil?: string
  ton?: string
  tgk?: string
  prs?: string
  pus?: string
  jam?: string
  lav?: string
  kaz?: string
  glv?: string
  ber?: string
  mey?: string
  kin?: string
  nfr?: string
  lit?: string
  nno?: string
  nob?: string
  smi?: string
  dzo?: string
  arc?: string
  ckb?: string
  kir?: string
  hin?: string
  nrf?: string
  nep?: string
  aze?: string
  amh?: string
  niu?: string
  pol?: string
  zdj?: string
  run?: string
  cat?: string
  rar?: string
  fas?: string
  mkd?: string
  khm?: string
  afr?: string
  her?: string
  hgm?: string
  kwn?: string
  loz?: string
  ndo?: string
  isl?: string
  heb?: string
  nbl?: string
  nso?: string
  zul?: string
  ben?: string
  lua?: string
  ukr?: string
  cal?: string
  cha?: string
  pau?: string
  fao?: string
  bar?: string
  kat?: string
  bjz?: string
  mfe?: string
  mah?: string
  tet?: string
  mon?: string
  nau?: string
}

interface Translations {
  ara: translation
  ces: translation
  cym: translation
  deu: translation
  est: translation
  fin: translation
  fra: translation
  hrv?: translation
  hun: translation
  ita: translation
  jpn?: translation
  kor: translation
  nld: translation
  per?: translation
  pol: translation
  por: translation
  rus: translation
  slk: translation
  spa: translation
  swe: translation
  urd: translation
  zho?: translation
}

export interface translation {
  official: string
  common: string
}

export interface Demonyms {
  eng: Eng2
  fra?: Fra3
}

export interface Eng2 {
  f: string
  m: string
}

export interface Fra3 {
  f: string
  m: string
}

export interface Maps {
  googleMaps: string
  openStreetMaps: string
}

export interface Gini {
  '2017'?: number
  '2010'?: number
  '2018'?: number
  '2009'?: number
  '2019'?: number
  '2008'?: number
  '2012'?: number
  '2015'?: number
  '2016'?: number
  '2011'?: number
  '2013'?: number
  '2003'?: number
  '2014'?: number
  '1998'?: number
  '1999'?: number
  '2006'?: number
  '2004'?: number
  '2005'?: number
  '1992'?: number
}

export interface Car {
  signs?: string[]
  side: string
}

export interface Flags {
  png: string
  svg: string
}

export interface CoatOfArms {
  png?: string
  svg?: string
}

export interface CapitalInfo {
  latlng?: number[]
}

export interface PostalCode {
  format: string
  regex?: string
}